refactor: response utils class
