+++ src/main/java/fr/insee/eno/ws/controller/utils/ResponseUtils.java
@@ -1,4 +1,4 @@
-package fr.insee.eno.ws.controller;
+package fr.insee.eno.ws.controller.utils;
 
 import java.io.File;
 import java.nio.file.Files;
@@ -7,14 +7,17 @@ import org.springframework.http.HttpHeaders;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.servlet.mvc.method.annotation.StreamingResponseBody;
 
-public class ResponseUtil {
+public class ResponseUtils {
+
+	private ResponseUtils() {}
 	
-	public static ResponseEntity<StreamingResponseBody> generateResponseFromFile(File file)throws Exception {
+	public static ResponseEntity<StreamingResponseBody> generateResponseFromFile(File file) {
 
-		StreamingResponseBody stream = out -> out.write(Files.readAllBytes(file.toPath())) ;
+		StreamingResponseBody stream = out -> out.write(Files.readAllBytes(file.toPath()));
 
-		return  ResponseEntity.ok()
+		return ResponseEntity.ok()
 				.header(HttpHeaders.CONTENT_DISPOSITION, "attachment;filename=\""+file.getName()+"\"")
 				.body(stream);
 	}
+
 }
